storage: add .ssignore functionality
The storage server can now ignore all files that are expressed as
regular expressions in DEFAULT_SSIGNORE_PATH. The idea is (clearly)
stolen from .gitignore (because it works and is easy). The
implementation, however, differs. .ssignore processes only POSIX regex,
so think like you are grepping through the directory for files to be
selected. Note that the storage server only searches in the root folder
and does not recurse.